Update products pagination in collection page
